Add material class and update texture class
Add a materiaal class with boost serialisation; it will be used for fast loading of materials. Moreover, update the texture class to use a static map for mapping texture resources to diifferent clients which need them: this way only one texture will be uploaded when multiple clients need a given one.